requirements.txt
3==0.0.0
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
bitarray==2.9.2
boto==2.49.0
boto3==1.34.0
botocore==1.34.0
build==1.0.3
CacheControl==0.13.1
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crashtest==0.4.1
crispy-bootstrap5==2023.10
Cython==3.0.8
cytoolz==0.12.3
dateparser==1.0.0
distlib==0.3.8
distro==1.9.0
Django==4.2.7
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-froala-editor==4.1.4
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
dydx-v3-python==2.1.0
ecdsa==0.18.0
eth-account==0.5.9
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==3.0.0
eth_abi==5.0.0
fastjsonschema==2.19.1
filelock==3.13.1
frozenlist==1.4.1
google-serp-api==1.0.4
google_search_results==2.4.2
h11==0.14.0
hashring==1.5.1
hexbytes==0.3.1
httpcore==1.0.2
httpx==0.26.0
idna==3.6
installer==0.7.0
jaraco.classes==3.3.0
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyring==24.3.0
kombu==5.3.5
lru-dict==1.2.0
Markdown==3.5.1
memcache==0.12.0
more-itertools==10.1.0
mpmath==1.0.0
msgpack==1.0.7
multidict==6.0.4
nltk==3.8.1
numpy==1.26.2
openai==1.7.2
opencv-python==4.8.1.78
packaging==23.2
parsimonious==0.9.0
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psycopg2==2.9.9
ptyprocess==0.7.0
pycryptodome==3.19.1
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==1.7.1
pymemcache==4.0.0
pyproject-toml==0.0.10
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyunormalize==15.1.0
pywin32==306
pywin32-ctypes==0.2.2
rapidfuzz==3.6.1
redis==5.0.1
referencing==0.32.0
regex==2023.12.25
reportlab==4.0.7
requests==2.31.0
requests-toolbelt==1.0.0
rlp==4.0.0
rpds-py==0.16.2
s3transfer==0.9.0
setuptools==69.0.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==7.11.0
sympy==1.6
textblob==0.17.1
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.1
tqdm==4.66.1
trove-classifiers==2023.11.29
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
web3==5.31.4
websockets==12.0
wheel==0.42.0
whitenoise==6.6.0
yarl==1.9.4
